# Poor man's RigidBodyPlant caching