--- a/drake/multibody/rigid_body_plant/contact_results.cc
+++ b/drake/multibody/rigid_body_plant/contact_results.cc
@@ -17,9 +17,6 @@ const ContactInfo<T>& ContactResults<T>::get_contact_info(int i) const {
 }
 
 template <typename T>
-ContactResults<T>::ContactResults() : contacts_() {}
-
-template <typename T>
 void ContactResults<T>::Clear() {
   contacts_.clear();
 }