--- a/drake/systems/analysis/runge_kutta2_integrator.h
+++ b/drake/systems/analysis/runge_kutta2_integrator.h
@@ -57,7 +57,6 @@ template <class T>
 bool RungeKutta2Integrator<T>::DoStep(const T& dt) {
   // Find the continuous state xc within the Context, just once.
   auto context = IntegratorBase<T>::get_mutable_context();
-  VectorBase<T>* xc = context->get_mutable_continuous_state_vector();
 
   // TODO(sherm1) This should be calculating into the cache so that
   // Publish() doesn't have to recalculate if it wants to output derivatives.
@@ -67,6 +66,8 @@ bool RungeKutta2Integrator<T>::DoStep(const T& dt) {
   // First stage is an explicit Euler step:
   // xc(t+h) = xc(t) + dt * xcdot(t, xc(t), u(t))
   const auto& xcdot0 = derivs0_->get_vector();
+  // The next call invalidates entries dependent on the state vector.
+  VectorBase<T>* xc = context->get_mutable_continuous_state_vector();
   xc->PlusEqScaled(dt, xcdot0);  // xc += dt * xcdot0
   T t = IntegratorBase<T>::get_context().get_time() + dt;
   IntegratorBase<T>::get_mutable_context()->set_time(t);
@@ -75,6 +76,10 @@ bool RungeKutta2Integrator<T>::DoStep(const T& dt) {
   IntegratorBase<T>::get_system().EvalTimeDerivatives(
       *IntegratorBase<T>::get_mutable_context(), derivs1_.get());
   const auto& xcdot1 = derivs1_->get_vector();
+
+  // The next call is needed again in order to properly invalidate cached
+  // entries.
+  xc = context->get_mutable_continuous_state_vector();
 
   // TODO(sherm1) Use better operators when available.
   xc->PlusEqScaled(dt * 0.5, xcdot1);