Since we still don't have System 2.0 caching this PR implements a temporary fix that substantially speeds up simulations.
The solution implemented is a `RigidBodyPlantContext` that inherits from `LeafContext` and where we place a number of `mutable` expensive to compute quantities (the "cache" entries).

In addition the PR identifies some possible design issues regarding the proper notification to the context when states are changed. Take a look at `*_integrator.h` where these notifications are triggered by the calls to `context->get_mutable_continuous_state_vector()`.

Speed up results for Valkyrie with full urdf model: from 24.0 secs to 9.0 secs to simulate 1.0 secs.

Speed up results for Valkyrie without grids, only spheres and boxes at the feet and head: from 10.0 secs to 7.5 secs to simulate 5.0 secs.

Most of the time was spent by the RBT collision detection methods that were unnecessarily being called 3 times per time step (for Euler, 6 for RK2).

cc'ing @sherm1, @SeanCurtis-TRI, @siyuanfeng-tri  

<!-- Reviewable:start -->
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/robotlocomotion/drake/4249)
<!-- Reviewable:end -->