@amcastro-tri 2016-11-23
+@sherm1 for feature review please.

---

Review status: 0 of 11 files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/4249#-:-KXI_ywfg-uEeUnLMjcb:b-ec6d43)*
<!-- Sent from Reviewable.io -->