@sherm1 2016-11-24
@amcastro-tri, this seems like a reasonable workaround but will it still be worth the trouble when @seancurtis-tri's fix for the all-vs-all contact & collision groups goes in? My impression is that that eliminates most of the computation time so then this change wouldn't have a big effect.

I notice that you added invalidation code for state variables changes but not for input changes or time. Are all the cached values here time- and input-independent?

Please consider how this can be unit tested. Can you add some tests somewhere to see whether the invalidations are being propagated correctly and that invalid results can't accidentally be used?

---

Reviewed 7 of 11 files at r1.
Review status: 7 of 11 files reviewed at latest revision, 1 unresolved discussion.

---

*[drake/multibody/rigid_body_plant/rigid_body_plant_context.h, line 43 at r1](https://reviewable.io:443/reviews/robotlocomotion/drake/4249#-KXInQoWm-rTWZPJ5fBZ:-KXInQoWm-rTWZPJ5fB_:b-oy72dp) ([raw file](https://github.com/robotlocomotion/drake/blob/57f30e7b578574dd0668893c4c4f608185ff8e18/drake/multibody/rigid_body_plant/rigid_body_plant_context.h#L43)):*
> ```objc
>   RigidBodyPlantContext& operator=(RigidBodyPlantContext&& other) = delete;
> 
>  public:
> ```

Data members should be private.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/4249#-:-KXIp506m0HLQw4l1D6Z:b-g2muzi)*
<!-- Sent from Reviewable.io -->