@amcastro-tri 2016-11-28
@seancurtis-tri PR #4206 sped up the Valkyrie simulation from 24.0 secs to 15 secs.
With the changes in this PR + #4206 the simulation time goes down to 6 secs (before #4206 it was 9.0 secs).
Essentially `ComputeMaximumDepthCollisionPoints` is being called three time per time step.
I can wait until a general caching capability is in place to start using it in RBP. I guess this PR is at least good to demonstrate the issue.

---

Review status: 7 of 11 files reviewed at latest revision, 1 unresolved discussion.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/robotlocomotion/drake/4249#-:-KXfYGdhj-F5eoVGcKcN:b-1xzhrl)*
<!-- Sent from Reviewable.io -->