@jwnimmer-tri 2016-11-28
I spent about 60 seconds reviewing this to try to understand the strategy.  I agree with @sherm1 that bringing this PR to a merge-able state (making it correct and well-tested) is likely so much (throwaway) work that we should instead of just focus on the actual System 2.0 caching implementation, instead of this PR that appears (upon a _very_ quick read) to be entirely throwaway work.