@amcastro-tri 2016-11-28
I agree on that the work in this PR is all throwaway code and cleaning it up for merging would incur in a waste of my time. 
I created #4270 to track performance improvements in our dynamics engine.